Split Trap Spaces into two implementations.
